benchmark: add rpmem_flush_drain and rpmem_mixed
- rpmem_flush_drain - performs n x rpmem_flush() [+ rpmem_drain()]
where n >= 0. If n == -1 there will be no rpmem_drain()
- rpmem_mixed - performs sequence of rpmem_* commands in single
worker thread operation
