+++ src/benchmarks/Makefile
@@ -149,7 +149,7 @@ LDFLAGS += -fsanitize=$(SANITIZE)
 endif
 
 ifeq ($(BUILD_RPMEM),y)
-SRC += rpmem_persist.cpp
+SRC += rpmem.cpp
 LIBS += -lrpmem $(shell pkg-config --libs libfabric)
 CONFIGS += pmembench_rpmem
 CXXFLAGS += -DRPMEM_AVAILABLE
