+++ src/benchmarks/pmembench_rpmem.cfg
@@ -8,18 +8,21 @@
 group = rpmem
 file = testfile.poolset.rpmem
 ops-per-thread = 10000
+threads = 32
+mem-mode = rand-wrap
 
 [rpmem_persist_DS8]
 bench = rpmem_persist
-threads = 1:+1:32
+persist-relaxed = true
 data-size = 8
 
-[rpmem_persist_DS512]
-bench = rpmem_persist
-threads = 1:+1:32
+[rpmem_flush_drain_DS512]
+bench = rpmem_flush_drain
+flushes-per-drain = 10
 data-size = 512
 
-[rpmem_persist_DS1024]
-bench = rpmem_persist
-threads = 1:+1:32
+[rpmem_mixed_DS1024]
+bench = rpmem_mixed
+# rpmem_persist() + RPMEM_PERSIST_RELAXED, rpmem_flush(), rpmem_drain() and rpmem_persist()
+workload = rfdp
 data-size = 1024
